Time: 54 ms (35.78%), Space: 31.9 MB (24.77%) - LeetHub

+++ 0264-ugly-number-ii/0264-ugly-number-ii.cpp
@@ -1,19 +1,19 @@
 class Solution {
 public:
     int nthUglyNumber(int n) {
-        set<long long> st;
+        set<long> st;
         st.insert(1);
-        int currentUgly = 1;
+        long currentUgly = 1;
         for(int i = 0; i < n; i++){
             currentUgly = *st.begin();
             
-            st.insert(currentUgly * 1LL * 2);
-            st.insert(currentUgly * 1LL * 3);
-            st.insert(currentUgly * 1LL * 5);
+            st.insert(currentUgly * 1L * 2);
+            st.insert(currentUgly * 1L * 3);
+            st.insert(currentUgly * 1L * 5);
             
             st.erase(currentUgly);
         }
         
-        return currentUgly;
+        return (int)currentUgly;
     }
 };
